testing class

+++ Server/services/app/migrations/20221101143712-create-transaction.js
@@ -26,7 +26,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       rating: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       testimoni: {
         type: Sequelize.STRING
